Fixing Coin Images
CoinMarketCap decided to change their API on me. The id
field in the ticker API is no longer the real ID. It is
actually called the "slug". This broke image grabbing
because their images are now hosted behind the real ID which
comes from the quick_search rather than directly from the ticker.

I don't know why they did this. But it is really annoying and it
hurts the performance of the app because I now have to do 2 network
calls rather than one.

These changes add a new call to the quick_search api from CMC which
maps the slugs to the IDs and then adds the real ID to the coin objects.